refactor(engine): extract models::download
The generic-HTTP concerns the ticket calls out — retry budgets, jitter,
Retry-After clamping — plus mirror resolution, SHA-256 verification and the
parallel download pool land in one file: model_mirror/apply_mirror,
download_verified and its retry/backoff machinery (AttemptFailure,
TrackedStream, backoff_delay, jitter_fraction), write_verified/verify_sha256,
cleanup_legacy/cleanup_orphan_staging, and the per-model download_{vad,tts,
diarize,manifest} entry points.

Pure move: paths.rs supplies cache_dir/ModelKind/is_cached*/fluidaudio_*
locations via `use super::paths::*`, and the two ANE staging functions
download_tts still calls (stage_ane_kokoro_voices,
stage_fluidaudio_kokoro_assets) stay put in mod.rs for now — they move with
the rest of staging in the next commit — reached back via `use super::{..}`.
write_verified/staging_path/verify_sha256 go `pub(super)` for
manifest_tests's five write_verified/verify_sha256 cases, which reach back
into models::download by name same as they reached into the shared mod.rs
before the split.

mod.rs's leftover `use` block needed cfg-gating it didn't need before: once
the download engine moved out, fs/io/Path/PathBuf/the anyhow Context trait/
with_stderr are only reached by the system_kokoro-macos-aarch64 staging code
still living here (plus its `cfg(test)` escape hatches), so an ONNX-only
build now finds them unused unless the import carries the same cfg as its
only remaining caller.

Refs #950
